GlController

+++ src/blocks/Block.hpp
@@ -5,7 +5,9 @@
 #include <string>
 
 struct block {
-	uint8_t id = 0;
+	uint16_t id = 0;
+
+	
 };
 
 #endif // !CHUNK_HPP
